requirements.txt
django==1.9.1
dj-database-url==0.3.0
djangorestframework==3.3.2
django-filter==0.12.0
Markdown==2.6.5
gunicorn==19.3.0
whitenoise==2.0.6
